+++ app/layout.tsx
@@ -1,34 +1,17 @@
-"use client";
-
 import "@/styles/globals.css";
-import Header from "@/app/components/Header";
-import { usePathname } from "next/navigation";
 import StyledComponentsRegistry from "@/lib/StyledComponentsRegistry";
-import { useGetHeaderTitle } from "@/utils/getHeaderTitle";
-import {
-  ContentContainer,
-  RootContainer,
-} from "./components/RootContainerStyled";
+import LayoutClient from "./components/LayoutClient";
 
 export default function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  const pathname = usePathname();
-
-  const title = useGetHeaderTitle(pathname);
-
-  const isBackType = title !== "";
-
   return (
     <html lang="en">
       <body>
         <StyledComponentsRegistry>
-          <RootContainer>
-            <Header type={isBackType ? "back" : "default"} title={title} />
-            <ContentContainer>{children}</ContentContainer>
-          </RootContainer>
+          <LayoutClient> {children} </LayoutClient>
         </StyledComponentsRegistry>
       </body>
     </html>
